Restoring a working state in main
[TODO]
Reintegrate functions in main

+++ components/mqtt_controller/include/thingsboard_types.h
@@ -4,7 +4,7 @@
 
 typedef struct {
     char* uri;
-    uint8_t port;
+    uint16_t port;
 } thingsboard_address_t;
 typedef struct {
     char* certificate;
